Create  Institute of Bioinformatics
Having learned that DNA is not a random string, newly admitted students of the computer science group at the Institute of Bioinformatics proposed using a compression algorithm that compresses repeating characters in a string.

The encoding is done as follows:
s = 'aaaabbсaa' is converted to 'a4b2с1a2', that is, groups of identical characters in the original string are replaced by this character and the number of times it is repeated at this position in the string.

Write a program that reads a string, encodes it using the proposed algorithm, and outputs the encoded sequence to standard output. The encoding must be case-sensitive.